Delete duplicate DriverProfile and update thr correct one.
